+++ lib/ooxml_parser/docx_parser/docx_data/document_structure/styles/document_defaults/paragraph_properties_default.rb
@@ -1,0 +1,20 @@
+module OoxmlParser
+  # Class for parsing `w:pPrDefault` tags
+  class ParagraphPropertiesDefault < OOXMLDocumentObject
+    # @return [ParagraphProperties] properties of run
+    attr_accessor :paragraph_properties
+
+    # Parse ParagraphPropertiesDefault object
+    # @param node [Nokogiri::XML:Element] node to parse
+    # @return [ParagraphPropertiesDefault] result of parsing
+    def parse(node)
+      node.xpath('*').each do |node_child|
+        case node_child.name
+        when 'pPr'
+          @paragraph_properties = ParagraphProperties.new(parent: self).parse(node_child)
+        end
+      end
+      self
+    end
+  end
+end